Recents: Delay cleanup of remaining tasks when dismissing
SystemServicesProxy.removeAllUserTask() can take upwards of 2000ms
to execute which causes the animation of tasks removing to be delayed.
Post a delayed runnable to handle the cleanup after the animations.

This patch also removes posting the dismissal of tasks in a runnable
as this code is already being executed on the UI thread from the
clear all recents onClick().

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java

Change-Id: I3ec2ac2cc2d7340b537b7b6b2d8f8459403cf4ee

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java